refactor(git): split github.go into per-domain files
Splits the 530-LOC git/github.go into five files grouped by GitHub
domain, with no behavior changes:

- github_clone.go        CloneGithub
- github_branch.go       CreateGithubBranch, DeleteGithubBranch
- github_pull_request.go CreateGithubPullRequest
- github_issue.go        CreateGithubIssue
- github_commit.go       AddFile/Folder/FilesToGithubBranch

Also moves the package-level `workDir` var into container.go (its
natural home alongside the container helper).

`dagger functions -m ./git` lists the same 11 functions with the
same descriptions as before the split. `go build`, `go vet` and
`gofmt -l` all clean.

Refs #262 (P3 — split oversized files; `git/github.go` slice).
